Refactor get() method in Field struct to accept a pointer to type T
The get() method in the Field struct has been modified to accept a pointer to type T instead of a value. This change allows for more flexibility when accessing the field data. The method has been updated to correctly calculate the memory offset based on the pointer. This refactor improves the overall usability and maintainability of the code.